+++ src/main.cpp
@@ -27,6 +27,7 @@
 #include "utils/cHandleArgument.h"
 #include "utils/common.h"
 #include "gameobjects/units/cUnits.h"
+#include "gameobjects/cUpgradeInfo.h"
 #include <string>
 
 #include <iostream>
@@ -41,6 +42,7 @@ std::shared_ptr<Graphics>   gfxdata;
 s_StructureInfo    			sStructureInfo[MAX_STRUCTURETYPES];
 s_UnitInfo         			sUnitInfo[MAX_UNITTYPES];
 s_UpgradeInfo               sUpgradeInfo[MAX_UPGRADETYPES];
+
 //s_SpecialInfo             sSpecialInfo[MAX_SPECIALTYPES];
 
 cSpecialInfos               specialInfos;
